+++ aesara/gpuarray/blocksparse.py
@@ -11,8 +11,8 @@ from aesara.gpuarray.basic_ops import (
 from aesara.gpuarray.type import gpu_context_type
 from aesara.gradient import grad_undefined
 from aesara.graph.basic import Apply
-from aesara.graph.params_type import ParamsType
 from aesara.link.c.op import _NoPythonExternalCOp
+from aesara.link.c.params_type import ParamsType
 from aesara.scalar import bool as bool_t
 from aesara.tensor import as_tensor_variable
 from aesara.tensor.type import discrete_dtypes
