+++ aesara/sandbox/fourier.py
@@ -12,7 +12,7 @@ import numpy.fft
 
 from aesara.graph.basic import Apply
 from aesara.graph.op import Op
-from aesara.graph.type import generic
+from aesara.link.c.type import generic
 from aesara.tensor.basic import as_tensor
 from aesara.tensor.type import zmatrix
