+++ aesara/tensor/extra_ops.py
@@ -12,9 +12,9 @@ from aesara.gradient import (
 )
 from aesara.graph.basic import Apply, Variable, equal_computations
 from aesara.graph.op import Op
-from aesara.graph.params_type import ParamsType
-from aesara.graph.type import EnumList, Generic
 from aesara.link.c.op import COp
+from aesara.link.c.params_type import ParamsType
+from aesara.link.c.type import EnumList, Generic
 from aesara.misc.safe_asarray import _asarray
 from aesara.raise_op import Assert
 from aesara.scalar import int32 as int_t
