+++ aesara/tensor/nnet/corr3d.py
@@ -6,9 +6,9 @@ import aesara
 from aesara.configdefaults import config
 from aesara.graph.basic import Apply
 from aesara.graph.op import _NoPythonOp
-from aesara.graph.params_type import ParamsType
-from aesara.graph.type import EnumList
 from aesara.link.c.op import OpenMPOp
+from aesara.link.c.params_type import ParamsType
+from aesara.link.c.type import EnumList
 from aesara.scalar import int64
 from aesara.tensor import blas_headers
 from aesara.tensor.basic import as_tensor_variable
