+++ tests/link/c/c_code/test_cenum.h
