+++ tests/link/c/c_code/test_quadratic_function.c
